tried to fix it
It tracking and messaging at the same time does not work. I have no idea why. The only thing, which I did not check, is that messages get lost during the transmission, but in theory it should be handled (bc we do not use OSM). Also it seems, that the module hangs up (does not execute its main loop further).